pyproject.toml
[project]
name = "moji"
version = "0.1.0"
description = "🍑 Multimodal RAG engine for sticker search"
requires-python = ">=3.11"
dependencies = [
    "google-generativeai>=0.8.0",
    "chromadb>=0.5.0",
    "fastapi>=0.115.0",
    "uvicorn>=0.32.0",
    "python-dotenv>=1.0.0",
    "Pillow>=10.0.0",
    "pydantic>=2.0.0",
]

[project.scripts]
moji = "moji.cli:main"

[build-system]
requires = ["setuptools>=68.0"]
build-backend = "setuptools.backends._legacy:_Backend"

[tool.setuptools.packages.find]
include = ["moji*"]
